console: Refactor connection profiles store
