+++ pkg/webui/console/store/reducers/connection-profiles.js
@@ -1,0 +1,46 @@
+// Copyright © 2024 The Things Network Foundation, The Things Industries B.V.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import {
+  GET_CONNECTION_PROFILES_LIST_SUCCESS,
+  DELETE_CONNECTION_PROFILE_SUCCESS,
+} from '@console/store/actions/connection-profiles'
+
+const defaultState = {
+  profiles: [],
+  profilesTotalCount: undefined,
+}
+
+const connectionProfiles = (state = defaultState, action) => {
+  const { type, payload } = action
+
+  switch (type) {
+    case GET_CONNECTION_PROFILES_LIST_SUCCESS:
+      return {
+        ...state,
+        profiles: payload.entities,
+        profilesTotalCount: payload.profilesTotalCount,
+      }
+    case DELETE_CONNECTION_PROFILE_SUCCESS:
+      return {
+        ...state,
+        profiles: state.profiles.filter(profile => profile.id !== payload.id),
+        profilesTotalCount: state.profilesTotalCount - 1,
+      }
+    default:
+      return state
+  }
+}
+
+export default connectionProfiles
