+++ pkg/webui/console/store/selectors/gateways.js
@@ -96,8 +96,3 @@ export const selectGatewayStatistics = state => {
 
   return statistics.stats
 }
-
-// Connection profiles
-export const selectConnectionProfiles = state => selectGatewayStore(state).connectionProfiles
-export const selectConnectionProfilesTotalCount = state =>
-  selectGatewayStore(state).connectionProfilesTotalCount
